+++ cassandra-test-image/src/main/resources/generate_certificates.sh
@@ -61,7 +61,7 @@ KEYSTORE="cert/.keystore"
 ###############
 
 ## Generate self-signed CA
-openssl req -x509 -newkey rsa:2048 -nodes -days 1 -subj "/C=TE/ST=TEST/L=TEST/O=TEST/OU=TEST/CN=CA"\
+openssl req -x509 -newkey rsa:2048 -nodes -days 365 -subj "/C=TE/ST=TEST/L=TEST/O=TEST/OU=TEST/CN=CA"\
  -keyout "$CA_KEY" -out "$CA_CERT"
 
 ## Import self-signed CA certificate to truststore
@@ -79,8 +79,9 @@ openssl req -newkey rsa:2048 -nodes -subj "/C=TE/ST=TEST/L=TEST/O=TEST/OU=TEST/C
  -keyout "$USER_KEY" -out "$USER_CSR"
 
 # Sign user certificate
-openssl x509 -req -sha256 -days 1\
+openssl x509 -req -sha256 -days 365\
  -in "$USER_CSR" -out "$USER_CERT"\
+ -extfile v3.ext\
  -CA "$CA_CERT" -CAkey "$CA_KEY" -CAcreateserial
 
 ## Convert key/certificate to PKCS12 format
@@ -111,12 +112,23 @@ cat > v3.ext <<-EOF
 authorityKeyIdentifier=keyid,issuer
 basicConstraints=CA:FALSE
 keyUsage = digitalSignature, nonRepudiation, keyEncipherment, dataEncipherment
-extendedKeyUsage = serverAuth
+extendedKeyUsage = serverAuth, clientAuth
 subjectAltName = @alt_names
 
 [alt_names]
 DNS.1 = localhost
+DNS.2 = cassandra-seed-dc1-rack1-node1
+DNS.3 = cassandra-seed-dc2-rack1-node1
+DNS.4 = cassandra-node-dc1-rack1-node2
+DNS.5 = cassandra-node-dc2-rack1-node2
 IP.1 = 127.0.0.1
+IP.2 = 172.29.0.2
+IP.3 = 172.29.0.3
+IP.4 = 172.29.0.4
+IP.5 = 172.29.0.5
+IP.6 = 172.29.0.6
+IP.7 = 172.29.0.7
+IP.8 = 172.29.0.8
 
 EOF